+++ doc/examples/aseprite/lib/main.dart
@@ -1,4 +1,3 @@
-import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/sprite_animation.dart';
 import 'package:flame/extensions/vector2.dart';
@@ -6,20 +5,14 @@ import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  final Vector2 size = await Flame.util.initialDimensions();
   runApp(
     GameWidget(
-      game: MyGame(size),
+      game: MyGame(),
     ),
   );
 }
 
 class MyGame extends BaseGame {
-  MyGame(Vector2 screenSize) {
-    size = screenSize;
-  }
-
   @override
   Future<void> onLoad() async {
     final image = await images.load('chopper.png');