+++ lib/components/base_component.dart
@@ -103,15 +103,25 @@ abstract class BaseComponent extends Component {
   List<ComponentEffect> get effects => _effectsHandler.effects;
 
   /// Uses the game passed in, or uses the game from [HasGameRef] otherwise,
+
   /// to prepare the child component before it is added to the list of children.
   /// Note that this component needs to be added to the game first if
   /// [this.gameRef] should be used to prepare the child.
   /// For children that don't need preparation from the game instance can
   /// disregard both the options given above.
-  void addChild(Component c, {Game gameRef}) {
+  Future<void> addChild(Component c, {Game gameRef}) async {
+    assert(
+      gameRef != null || this is HasGameRef,
+      "Need gameRef either as an argument or from the HasGameRef mixin",
+    );
     gameRef ??= (this as HasGameRef).gameRef;
     if (gameRef is BaseGame) {
       gameRef.prepare(c);
+    }
+
+    final childOnLoadFuture = c.onLoad();
+    if (childOnLoadFuture != null) {
+      await childOnLoadFuture;
     }
     _children.add(c);
   }
@@ -123,6 +133,8 @@ abstract class BaseComponent extends Component {
   void clearChildren() {
     _children.clear();
   }
+
+  bool containsChild(Component c) => _children.contains(c);
 
   /// This method first calls the passed handler on the leaves in the tree, the children without any children of their own.
   /// Then it continues through all other children.